--- activitysim/defaults/__init__.py
@@ -1,3 +1,0 @@
-# ActivitySim
-# Copyright (C) 2014-2015 Synthicity, LLC
-# See full license in LICENSE.txt.
