+++ example/activitysim.py
@@ -1,0 +1,44 @@
+import urbansim.sim.simulation as sim
+import urbansim.utils.misc as usim_misc
+import os
+from activitysim import activitysim as asim
+import openmatrix as omx
+from activitysim import skim
+import numpy as np
+import pandas as pd
+
+
+"""
+Definition of terms:
+
+CDAP = coordinated daily activity pattern
+TDD = tour departure and duration
+"""
+
+
+warnings.filterwarnings('ignore', category=pd.io.pytables.PerformanceWarning)
+pd.options.mode.chained_assignment = None
+
+
+@sim.injectable(cache=True)
+def settings():
+    with open(os.path.join(misc.configs_dir(), "settings.yaml")) as f:
+        settings = yaml.load(f)
+        # monkey patch on the settings object since it's pretty global
+        # but will also be available as injectable
+        sim.settings = settings
+        return settings
+
+
+@sim.injectable(cache=True)
+def store(settings):
+    return pd.HDFStore(
+        os.path.join(misc.data_dir(), settings["store"]),
+        mode='r')
+
+
+# these are the alternatives for the workplace choice
+@sim.table()
+def zones():
+    # I grant this is a weird idiom but it helps to name the index
+    return pd.DataFrame({"TAZ": np.arange(1454)+1}).set_index("TAZ")
