+++ example/models/destination.py
@@ -1,0 +1,73 @@
+"""
+Given the tour generation from the above, each tour needs to have a
+destination, so in this case tours are the choosers (with the associated
+person that's making the tour)
+"""
+
+
+@sim.table()
+def destination_choice_size_terms():
+    f = os.path.join('configs', 'destination_choice_size_terms.csv')
+    return pd.read_csv(f)
+
+
+@sim.table()
+def destination_choice_spec():
+    f = os.path.join('configs', 'destination_choice_alternatives_sample.csv')
+    return asim.read_model_spec(f, stack=False).head(5)
+
+
+@sim.model()
+def destination_choice(non_mandatory_tours,
+                       persons,
+                       households,
+                       land_use,
+                       zones,
+                       distance_skim,
+                       destination_choice_spec):
+
+    tours = non_mandatory_tours
+
+    # FIXME these models don't have size terms at the moment
+
+    # FIXME these models don't use stratified sampling
+
+    # FIXME is the distance to the second trip based on the choice of the
+    # FIXME first trip - ouch!
+
+    # choosers are tours - in a sense tours are choosing their destination
+    choosers = sim.merge_tables(tours.name, tables=[tours,
+                                                    persons,
+                                                    households])
+
+    skims = {
+        "distance": distance_skim
+    }
+
+    choices_list = []
+    # segment by trip type and pick the right spec for each person type
+    for name, segment in choosers.groupby('tour_type'):
+
+        # FIXME - there are two options here escort with kids and without
+        if name == "escort":
+            continue
+
+        print "Running segment '%s' of size %d" % (name, len(segment))
+
+        choices, _ = \
+            asim.simple_simulate(segment,
+                                 zones.to_frame(),
+                                 destination_choice_spec[name],
+                                 skims,
+                                 skim_join_name="TAZ",
+                                 mult_by_alt_col=False,
+                                 sample_size=50)
+
+        choices_list.append(choices)
+
+    choices = pd.concat(choices_list)
+
+    print "Choices:\n", choices.describe()
+    # every trip now has a destination which is the index from the
+    # alternatives table - in this case it's the destination taz
+    sim.add_column("non_mandatory_tours", "destination", choices)
