+++ example/tables/accessibility.py
@@ -1,0 +1,8 @@
+@sim.table(cache=True)
+def accessibility(store):
+    df = store["skims/accessibility"]
+    df.columns = [c.upper() for c in df.columns]
+    return df
+
+
+sim.broadcast('accessibility', 'households', cast_index=True, onto_on='TAZ')
